feat(VForm): port to v3
